Remove less interested where_clause_factory
I'd not prefer to allocate its less interested object, it doesn't
improve readability, and make code slower.
